+++ Assets/Prefabs/Character/First Person Player.prefab
@@ -358,7 +358,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1593727746557118}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: 1.4880002, z: 0}
+  m_LocalPosition: {x: -368, y: 11.838, z: -238}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 4207293392497186}
@@ -562,6 +562,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   speed: 5
   canRun: 0
+  canMove: 1
   runSpeed: 9
   runningKey: 304
   cameraBobbingReference: {fileID: 612920248}
@@ -614,6 +615,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   maxHealth: 0
   currentHealth: 0
+  currentWeapon: {fileID: 7783858057294080571}
+  firstPersonCamera: {fileID: 20437472682476688}
 --- !u!1 &1963958859065967600
 GameObject:
   m_ObjectHideFlags: 0
@@ -857,10 +860,10 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   character: {fileID: 4207293392497186}
-  sensitivity: 200
-  smoothing: 2
   fpsCamera: {fileID: 20437472682476688}
   firstPersonMovement: {fileID: 114963311885979670}
+  sensitivity: 200
+  smoothing: 2
 --- !u!114 &612920248
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1638,3 +1641,14 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 2744828046676816823, guid: dfaccb852e93c7940b9352bef0f7f8a0, type: 3}
   m_PrefabInstance: {fileID: 1417636215970327264}
   m_PrefabAsset: {fileID: 0}
+--- !u!114 &7783858057294080571 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 9199065172871774427, guid: dfaccb852e93c7940b9352bef0f7f8a0, type: 3}
+  m_PrefabInstance: {fileID: 1417636215970327264}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fe1086053887521468bf7ab4d81e567a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
