+++ Assets/Prefabs/Enemies/Alien.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3562cd2d09809964e96064e736630072
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
