--- Assets/Scripts/Interfaces/ICanAttack.cs
@@ -1,8 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public interface ICanAttack {
-    void AttackWithWeapon();
-    IEnumerator StartReload();
-}
