--- Assets/Scripts/Interfaces/ICanAttack.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 425d615723066054b83c41523ca23f84
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
