+++ UserSettings/EditorUserSettings.asset
@@ -14,6 +14,9 @@ EditorUserSettings:
     RecentlyUsedScenePath-2:
       value: 22424703114646680e0b0227036c6c111b07142f1f2b233e2867083debf42d
       flags: 0
+    RecentlyUsedScenePath-3:
+      value: 22424703114646680e0b0227036c7315001214157d1712353e3d5326ece92021
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
